+++ exasol_integration_test_docker_environment/test/test_api_external_test_environment.py
@@ -1,7 +1,7 @@
 import unittest
 from sys import stderr
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
