+++ exasol_integration_test_docker_environment/test/test_api_logging.py
@@ -13,7 +13,7 @@ from typing import (
     Optional,
 )
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
