+++ exasol_integration_test_docker_environment/test/test_base_task.py
@@ -8,7 +8,7 @@ from luigi import (
     Parameter,
 )
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
 from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
