+++ exasol_integration_test_docker_environment/test/test_base_task_cleanup.py
@@ -7,7 +7,7 @@ from luigi import (
     IntParameter,
 )
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
 from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
