+++ exasol_integration_test_docker_environment/test/test_certificate_docker_build.py
@@ -1,10 +1,9 @@
-import shutil
 import unittest
 
 import luigi
 
 import exasol_integration_test_docker_environment.certificate_resources.container
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
 from exasol_integration_test_docker_environment.lib.config.docker_config import (
