+++ exasol_integration_test_docker_environment/test/test_common_run_task.py
@@ -2,7 +2,7 @@ import unittest
 
 import luigi
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
