+++ exasol_integration_test_docker_environment/test/test_generate_graph_plot.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import luigi
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
