+++ exasol_integration_test_docker_environment/test/test_test_env_reuse.py
@@ -5,7 +5,7 @@ import luigi
 from exasol_integration_test_docker_environment.cli.options import (
     test_environment_options,
 )
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
 from exasol_integration_test_docker_environment.lib.config.docker_config import (
