+++ exasol_integration_test_docker_environment/testing/luigi_utils.py
@@ -2,7 +2,7 @@ import shutil
 
 import luigi
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
 from exasol_integration_test_docker_environment.lib.docker.images.clean.clean_images import (
