+++ test/unit/test_env_variable.py
@@ -5,7 +5,7 @@ from unittest import mock
 import luigi
 import pytest
 
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
