+++ test/unit/test_termination_handler.py
@@ -14,7 +14,7 @@ import pytest
 from exasol_integration_test_docker_environment.cli.termination_handler import (
     TerminationHandler,
 )
-from exasol_integration_test_docker_environment.lib.api.run_task import (
+from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
     run_task,
 )
